kill warming
git-svn-id: http://192.168.20.11/sequoiadb/trunk@16648 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/client/common.c
@@ -273,14 +273,14 @@ INT32 clientCheckRetMsgHeader( const CHAR *pSendBuf, const CHAR *pRecvBuf )
 {
    INT32 rc = SDB_OK ;
    INT32 sendOpCode = 0 ;
-   INT32 recvOpCode = 0 ;
+   UINT32 recvOpCode = 0 ;
    if ( NULL == pSendBuf || NULL == pRecvBuf )
    {
       rc = SDB_INVALIDARG ;
       goto done ;
    }
    sendOpCode = ((MsgHeader*)pSendBuf)->opCode ;
-   recvOpCode = ((MsgHeader*)pRecvBuf)->opCode ;
+   recvOpCode = (UINT32)(((MsgHeader*)pRecvBuf)->opCode) ;
    if ( MAKE_REPLY_TYPE( sendOpCode ) != recvOpCode )
    {
 //      rc = SDB_UNEXPECTED_RESULT ;
